+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## [Unreleased]
+
+## [2.6.0] - 2025-03-06
 - Edit `AccountGuide.settingsFilled`
 - Add `MyPreferences.guideHide`
 - Add `Account.guide` endpoint and `AccountGuide` resource